sidebarusing streamlit/sidebarusing streamlit/requirements.txt
streamlit
streamlit_option_menu
psycopg2-binary